fix: list use array to list municipalities of comma-separated provinces
