+++ app/src/scripts/by_region.js
@@ -2,7 +2,7 @@ const path = require('path')
 const prompt = require('../lib/prompt')
 const { formatDisplay } = require('../lib/format_display')
 const selectDataSource = require('../lib/selector')
-const regions = require('../../data/regions.json')
+const regions = require('../../config/regions.json')
 
 // Asks for a prompt to enter a region name.
 // Lists all municipalities under the provinces of a region.
